Attach option-specified hints to the root transform.
